+++ universal-login-react/src/ui/UFlow/ApproveDevice.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const ApproveDevice = () => {
+  return (
+    <div>ApproveDevice</div>
+  );
+};
