+++ universal-login-react/src/ui/UFlow/TokenAssets.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import {TokenDetailsWithBalance} from '@universal-login/commons';
+import {TokenAsset} from './TokenAsset';
+
+interface TokenAssetsProps {
+  tokensDetailsWithBalance: TokenDetailsWithBalance[];
+}
+
+export const TokenAssets = ({tokensDetailsWithBalance}: TokenAssetsProps) => {
+  return (
+    <div>
+      {tokensDetailsWithBalance.map((token: TokenDetailsWithBalance) => (
+        <TokenAsset token={token} key={token.symbol}/>
+      ))}
+    </div>
+  );
+};
