+++ universal-login-react/src/ui/UFlow/USettings.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const USettings = () => {
+  return (
+    <div>Settings</div>
+  );
+};
